+++ inlang/source-code/rpc/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @inlang/rpc
 
+## 0.2.0
+
+### Minor Changes
+
+- b7dfc781e: change message format match from object to array
+
+### Patch Changes
+
+- Updated dependencies [b7dfc781e]
+  - @inlang/message@2.0.0
+  - @inlang/sdk@0.8.0
+
 ## 0.1.0
 
 ### Minor Changes